+++ CHANGELOG.md
@@ -2315,6 +2315,14 @@ Alternately, the _Daily Diary of my Programming Dreams_.
 * **angular Cli**
 
     1.  Added [fiddles/angular2-cli/fiddle-0037-EntAgGridStateManagement](fiddles/angular2-cli/fiddle-0037-EntAgGridStateManagement)
+    2.  Added [fiddles/angular2-cli/fiddle-0039-NotificationsHW](fiddles/angular2-cli/fiddle-0039-NotificationsHW)
+    3.  Added [fiddles/angular2-cli/fiddle-0039-WebNotificationsHW](fiddles/angular2-cli/fiddle-0039-WebNotificationsHW)
 
-* Added [fiddles/angular2-cli/fiddle-0039-NotificationsHW](fiddles/angular2-cli/fiddle-0039-NotificationsHW)
-* Added [fiddles/angular2-cli/fiddle-0039-WebNotificationsHW](fiddles/angular2-cli/fiddle-0039-WebNotificationsHW)
+
+### 360_python_11 ~ [Issue 360](https://github.com/bradyhouse/house/issues/360)
+
+* **python**
+
+    1.  Added [fiddles/python/fiddle-0011-Mp4TxtConverter](fiddles/python/fiddle-0011-Mp4TxtConverter)
+
+